# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  models.Publication:
    properties:
      completion_date:
        type: string
      creation_date:
        type: string
      description:
        description: Краткое описание события
        type: string
      formation_date:
        type: string
      id:
        type: integer
      moderator_id:
        type: integer
      status:
        description: Установлен статус по умолчанию
        type: string
      title:
        description: Название публикации (исторического события)
        type: string
      trust_score:
        type: number
      user_id:
        type: integer
    type: object
...
  models.UpdatePublicationDTO:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API для управления событиями, публикациями и изображениями.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: RIP API
  version: "1.0"
paths:
...
  /publications:
    get:
      consumes:
      - application/json
      description: Возвращает список публикаций с возможностью фильтрации по статусу
        и диапазону дат.
      parameters:
      - description: Статус публикации
        in: query
        name: status
        type: string
      - description: 'Дата начала (формат: yyyy-mm-dd)'
        in: query
        name: startDate
        type: string
      - description: 'Дата окончания (формат: yyyy-mm-dd)'
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список публикаций
          schema:
            items:
              $ref: '#/definitions/models.Publication'
            type: array
        "400":
          description: Неверный формат параметров
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Публикации не найдены
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить список публикаций
      tags:
      - Publications
  /publications/{publicationID}:
    delete:
      consumes:
      - application/json
      description: Удаляет публикацию по уникальному идентификатору.
      parameters:
      - description: Идентификатор публикации
        in: path
        name: publicationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Публикация успешно удалена
        "400":
          description: Неверный формат ID публикации
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Публикация не найдена
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Удалить публикацию по ID
      tags:
      - Publications
    get:
      consumes:
      - application/json
      description: Возвращает публикацию по уникальному идентификатору.
      parameters:
      - description: Идентификатор публикации
        in: path
        name: publicationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Публикация
          schema:
            $ref: '#/definitions/models.Publication'
        "400":
          description: Неверный формат ID публикации
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Публикация не найдена
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Получить публикацию по ID
      tags:
      - Publications
    put:
      consumes:
      - application/json
      description: Обновляет информацию о публикации по уникальному идентификатору.
      parameters:
      - description: Идентификатор публикации
        in: path
        name: publicationID
        required: true
        type: integer
      - description: Обновленные данные публикации
        in: body
        name: publication
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePublicationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная публикация
          schema:
            $ref: '#/definitions/models.Publication'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Публикация не найдена
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Обновить публикацию по ID
      tags:
      - Publications
  /publications/{publicationID}/finalize:
    post:
      consumes:
      - application/json
      description: Переводит публикацию в финальный статус.
      parameters:
      - description: Идентификатор публикации
        in: path
        name: publicationID
        required: true
        type: integer
      - description: Статус публикации
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Завершенная публикация
          schema:
            $ref: '#/definitions/models.Publication'
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Публикация не найдена
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Неверный статус для завершения публикации
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Завершить публикацию
      tags:
      - Publications
  /publications/{publicationID}/form:
    post:
      consumes:
      - application/json
      description: Переводит черновик публикации в статус работы.
      parameters:
      - description: Идентификатор публикации
        in: path
        name: publicationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Публикация в статусе работы
          schema:
            $ref: '#/definitions/models.Publication'
        "400":
          description: Неверный формат ID публикации
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Публикация не найдена
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Перевести публикацию в статус работы
      tags:
      - Publications
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Use the JWT Bearer Token for authorization. Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"